+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-08-03
+### Added
+- Support for importing from ~/.ssh/config
+### Fixed
+- Configuration file paths on Windows
+- Terminal cursor visibility
+
 ## [1.0.1] - 2025-07-29
 ### Changed
 - Table constraints